Install newer version of trpc/server and connect trpc with express
