+++ server/package.json
@@ -9,7 +9,7 @@
     "prettify": "prettier --loglevel warn --cache --write . -ext .json,.ts,.tsx,.js,.jsx,.css,.scss,.md,.mdx,.html --ignore-unknown"
   },
   "dependencies": {
-    "@trpc/server": "^10.45.2",
+    "@trpc/server": "^11.0.0-rc.467",
     "express": "^4.19.2"
   },
   "devDependencies": {
